requirements/base.txt
django-allauth
Django==5.1.7
django-autoslug==1.9.9
django-environ==0.12.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg
python-decouple
requests
django-cors-headers
social-auth-app-django
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis
celery
flower
django-celery-email
django-celery-beat
argh==0.26.2
PyYAML==5.3.1
watchdog
channels[daphne]
django-money==3.5.4
django-imagekit